sending test packets
