+++ .gitignore
@@ -1,2 +1,3 @@
 output
 install.log
+*.pyc
