+++ serial_rl.py
@@ -1,0 +1,58 @@
+#!/usr/bin/env python
+
+import serial
+import time
+
+class SerialRL:
+
+  CMD_GET_STATE = 1
+  CMD_GET_VERSION = 2
+  CMD_SET_CHANNEL = 3
+  CMD_GET_PACKET = 4
+  CMD_SEND_PACKET = 5
+
+  def __init__(self, serial_port):
+    self.ser = serial.Serial(serial_port, 19200, rtscts=1, timeout=1)
+    self.buf = bytearray()
+
+  def do_command(self, command, param=""): 
+    self.send_command(command, param)
+    return self.get_response()
+
+  def send_command(self, command, param=""): 
+    self.ser.write(chr(command))
+    if len(param) > 0:
+      self.ser.write(param)
+
+  def get_response(self, timeout=0): 
+    start = time.time()
+    while 1:
+      bytesToRead = self.ser.inWaiting()
+      if bytesToRead > 0:
+        self.buf.extend(self.ser.read(bytesToRead))
+      eop = self.buf.find(b'\x00',0)
+      if eop >= 0:
+        r = self.buf[:eop]
+        del self.buf[:(eop+1)]
+        return r
+      if (timeout > 0) and (start + timeout < time.time()):
+        return bytearray()
+      time.sleep(0.005)
+  
+  def sync(self):
+    while 1:
+      self.send_command(self.CMD_GET_STATE) 
+      data = self.get_response(1)
+      if data == "OK":
+        print "RileyLink " + data
+        break 
+      print "retry"
+ 
+    while 1:
+      self.send_command(self.CMD_GET_VERSION) 
+      data = self.get_response(1)
+      if len(data) >= 3:
+        print "Version: " + data
+        break 
+      print "retry"
+
